moved multiple functions to datatables
